fix(tests): remove old tests
