--- packages/core/tests-new/utils/index.ts
@@ -1,28 +1,0 @@
-import { CairoCustomEnum } from "starknet";
-
-export function sleep(ms: number) {
-  return new Promise((resolve) => setTimeout(resolve, ms));
-}
-
-export function generateRandomTokenId(): number {
-  return Math.floor(Math.random() * 10000) + 1;
-}
-
-type VariantKey = "Listing" | "Auction" | "Offer" | "CollectionOffer";
-
-export function getTypeFromCairoCustomEnum(cairoCustomEnum: CairoCustomEnum) {
-  const keyMap = {
-    Listing: "LISTING",
-    Auction: "AUCTION",
-    Offer: "OFFER",
-    CollectionOffer: "COLLECTION_OFFER"
-  };
-
-  for (const key in cairoCustomEnum.variant) {
-    if (cairoCustomEnum.variant[key as VariantKey] !== undefined) {
-      return keyMap[key as VariantKey] || "Unknown";
-    }
-  }
-
-  throw new Error("No valid variant found in CairoCustomEnum");
-}
