+++ packages/core/tests/createAccount.test.ts
@@ -1,4 +1,3 @@
-import { expect } from "chai";
 import { RpcProvider } from "starknet";
 
 import { createAccount } from "../src/actions/account/account";
@@ -14,12 +13,12 @@ describe("ArkProject Account Creation", () => {
     const accountDetails = await createAccount(provider);
 
     // Assertions to verify account creation
-    expect(accountDetails).to.have.property("address");
-    expect(accountDetails.address).to.be.a("string");
-    expect(accountDetails).to.have.property("privateKey");
-    expect(accountDetails.privateKey).to.be.a("string");
-    expect(accountDetails).to.have.property("publicKey");
-    expect(accountDetails.publicKey).to.be.a("string");
-    expect(accountDetails).to.have.property("account");
+    expect(accountDetails).toHaveProperty("address");
+    expect(typeof accountDetails.address).toBe("string");
+    expect(accountDetails).toHaveProperty("privateKey");
+    expect(typeof accountDetails.privateKey).toBe("string");
+    expect(accountDetails).toHaveProperty("publicKey");
+    expect(typeof accountDetails.publicKey).toBe("string");
+    expect(accountDetails).toHaveProperty("account");
   });
 });
