+++ packages/core/tests/fulfillOffer.test.ts
@@ -1,92 +1,219 @@
-import chai, { expect } from "chai";
-import chaiAsPromised from "chai-as-promised";
-import { RpcProvider, shortString } from "starknet";
+import { shortString } from "starknet";
 
+import { config } from "../examples/config";
 import {
+  STARKNET_ETH_ADDRESS,
+  STARKNET_NFT_ADDRESS
+} from "../examples/constants";
+import { getCurrentTokenId } from "../examples/utils/getCurrentTokenId";
+import { mintERC20 } from "../examples/utils/mintERC20";
+import { mintERC721 } from "../examples/utils/mintERC721";
+import { whitelistBroker } from "../examples/utils/whitelistBroker";
+import {
+  approveERC20,
+  approveERC721,
   createAccount,
   createOffer,
+  fetchOrCreateAccount,
   fulfillOffer,
   getOrderStatus,
   OfferV1
 } from "../src";
-import { generateRandomTokenId } from "./utils";
-
-chai.use(chaiAsPromised);
 
 describe("ArkProject Listing and Offer Fulfillment", () => {
   it("should create an offer and fulfill the offer", async function () {
-    this.timeout(10000); // Extend timeout if necessary
+    const { arkProvider, starknetProvider } = config;
 
-    // Initialize the RPC provider with the ArkChain node URL
-    const starknetProvider = new RpcProvider({
-      nodeUrl: "http://0.0.0.0:7777"
-    });
+    const solisAdminAccount = await fetchOrCreateAccount(
+      config.arkProvider,
+      process.env.SOLIS_ADMIN_ADDRESS_DEV,
+      process.env.SOLIS_ADMIN_PRIVATE_KEY_DEV
+    );
 
-    // Initialize the RPC provider with the katana node URL for starknet
-    const arkProvider = new RpcProvider({
-      nodeUrl: "http://0.0.0.0:7777"
-    });
+    await whitelistBroker(config, solisAdminAccount, 123);
 
     // Create a new account for the listing using the provider
     const { account: arkAccount } = await createAccount(arkProvider);
-    const { account: starknetAccount } = await createAccount(starknetProvider);
+
+    // Create a new account for fulfilling the offer
+    const starknetFulfillerAccount = await fetchOrCreateAccount(
+      config.starknetProvider,
+      process.env.STARKNET_ACCOUNT2_ADDRESS,
+      process.env.STARKNET_ACCOUNT2_PRIVATE_KEY
+    );
+
+    expect(starknetFulfillerAccount).toBeDefined();
+
+    await mintERC721(starknetProvider, starknetFulfillerAccount);
+
+    const tokenId = await getCurrentTokenId(config, STARKNET_NFT_ADDRESS);
 
     // Define the order details
     const order: OfferV1 = {
       brokerId: 123, // The broker ID
-      tokenAddress:
-        "0x01435498bf393da86b4733b9264a86b58a42b31f8d8b8ba309593e5c17847672", // The token address
-      tokenId: generateRandomTokenId(), // The ID of the token
+      tokenAddress: STARKNET_NFT_ADDRESS, // The token address
+      tokenId, // The ID of the token
       startAmount: 600000000000000000 // The starting amount for the order
     };
 
-    // Create the listing on the arkchain using the order details
-    const orderHash = await createOffer(
-      arkProvider,
-      starknetAccount,
-      arkAccount,
-      order
+    const starknetOffererAccount = await fetchOrCreateAccount(
+      starknetProvider,
+      process.env.STARKNET_ACCOUNT1_ADDRESS,
+      process.env.STARKNET_ACCOUNT1_PRIVATE_KEY
     );
-    expect(orderHash).to.exist;
+
+    await mintERC20(
+      starknetProvider,
+      starknetOffererAccount,
+      order.startAmount
+    );
+
+    // for allowance
+    await approveERC20(config, {
+      starknetAccount: starknetOffererAccount,
+      contractAddress: STARKNET_ETH_ADDRESS,
+      amount: order.startAmount
+    });
+
+    // Create the listing on the arkchain using the order details
+    const orderHash = await createOffer(config, {
+      starknetAccount: starknetOffererAccount,
+      arkAccount,
+      offer: order
+    });
+
+    expect(orderHash).toBeDefined();
+
+    await approveERC721(config, {
+      contractAddress: STARKNET_NFT_ADDRESS,
+      starknetAccount: starknetFulfillerAccount
+    });
 
     await new Promise((resolve) => setTimeout(resolve, 2000));
 
-    const { orderStatus: orderStatusBefore } = await getOrderStatus(
-      orderHash,
-      arkProvider
-    );
-    expect(shortString.decodeShortString(orderStatusBefore)).to.equal("OPEN");
-
-    // Create a new account for fulfilling the offer
-    const { account: starknetFulfillerAccount } =
-      await createAccount(starknetProvider);
-
-    expect(starknetFulfillerAccount).to.exist;
+    await expect(
+      getOrderStatus(config, { orderHash }).then((res) =>
+        shortString.decodeShortString(res.orderStatus)
+      )
+    ).resolves.toEqual("OPEN");
 
     // Define the fulfill details
-    const fulfill_info = {
-      order_hash: orderHash,
-      token_address: order.tokenAddress,
-      token_id: order.tokenId
+    const fulfillOfferInfo = {
+      orderHash,
+      tokenAddress: order.tokenAddress,
+      tokenId: order.tokenId,
+      brokerId: 123
     };
 
     // Fulfill the offer
-    await fulfillOffer(
-      arkProvider,
-      starknetFulfillerAccount,
+    await fulfillOffer(config, {
+      starknetAccount: starknetFulfillerAccount,
       arkAccount,
-      fulfill_info
-    );
+      fulfillOfferInfo
+    });
+
+    await new Promise((resolve) => setTimeout(resolve, 3000));
+    const { orderStatus: orderStatusBetween } = await getOrderStatus(config, {
+      orderHash
+    });
+    expect(shortString.decodeShortString(orderStatusBetween)).toBe("FULFILLED");
 
     await new Promise((resolve) => setTimeout(resolve, 5000));
+    const { orderStatus: orderStatusAfter } = await getOrderStatus(config, {
+      orderHash
+    });
+    expect(shortString.decodeShortString(orderStatusAfter)).toBe("EXECUTED");
+  }, 60000);
 
-    const { orderStatus: orderStatusAfter } = await getOrderStatus(
+  it("should create an offer and fail to fulfill the offer because owner of token changed", async function () {
+    const { arkProvider, starknetProvider } = config;
+
+    const solisAdminAccount = await fetchOrCreateAccount(
+      config.arkProvider,
+      process.env.SOLIS_ADMIN_ADDRESS_DEV,
+      process.env.SOLIS_ADMIN_PRIVATE_KEY_DEV
+    );
+
+    await whitelistBroker(config, solisAdminAccount, 123);
+
+    const { account: arkAccount } = await createAccount(arkProvider);
+    // Create a new account for the listing using the provider
+    const starknetFulfillerAccount = await fetchOrCreateAccount(
+      config.starknetProvider,
+      process.env.STARKNET_ACCOUNT2_ADDRESS,
+      process.env.STARKNET_ACCOUNT2_PRIVATE_KEY
+    );
+
+    await mintERC721(config.starknetProvider, starknetFulfillerAccount);
+    await approveERC721(config, {
+      contractAddress: STARKNET_NFT_ADDRESS,
+      starknetAccount: starknetFulfillerAccount
+    });
+
+    const tokenId = await getCurrentTokenId(config, STARKNET_NFT_ADDRESS);
+
+    // Create a new account for the listing using the provider
+    const starknetAccount = await fetchOrCreateAccount(
+      config.starknetProvider,
+      process.env.STARKNET_ACCOUNT1_ADDRESS,
+      process.env.STARKNET_ACCOUNT1_PRIVATE_KEY
+    );
+
+    // Define the order details
+    const order: OfferV1 = {
+      brokerId: 123, // The broker ID
+      tokenAddress: STARKNET_NFT_ADDRESS, // The token address
+      tokenId: tokenId, // The ID of the token
+      startAmount: 600000000000000000, // The starting amount for the order
+      currencyAddress: STARKNET_ETH_ADDRESS // The ERC20 address
+    };
+
+    await mintERC20(starknetProvider, starknetAccount, order.startAmount);
+
+    // for allowance
+    await approveERC20(config, {
+      starknetAccount: starknetAccount,
+      contractAddress: STARKNET_ETH_ADDRESS,
+      amount: order.startAmount
+    });
+
+    // Create the offer on the arkchain using the order details
+    const orderHash = await createOffer(config, {
+      starknetAccount,
+      arkAccount,
+      offer: order
+    });
+
+    expect(orderHash).toBeDefined();
+
+    await new Promise((resolve) => setTimeout(resolve, 2000));
+
+    const { orderStatus: orderStatusBefore } = await getOrderStatus(config, {
+      orderHash
+    });
+    expect(shortString.decodeShortString(orderStatusBefore)).toBe("OPEN");
+
+    expect(starknetFulfillerAccount).toBeDefined();
+
+    // Define the fulfill details
+    const fulfill_info = {
       orderHash,
-      arkProvider
-    );
+      tokenAddress: order.tokenAddress,
+      tokenId: order.tokenId,
+      brokerId: 123
+    };
 
-    expect(shortString.decodeShortString(orderStatusAfter)).to.equal(
-      "FULFILLED"
-    );
-  });
+    // Fulfill the offer
+    await fulfillOffer(config, {
+      starknetAccount: starknetFulfillerAccount,
+      arkAccount,
+      fulfillOfferInfo: fulfill_info
+    });
+
+    await new Promise((resolve) => setTimeout(resolve, 3000));
+    const { orderStatus: orderStatusBetween } = await getOrderStatus(config, {
+      orderHash
+    });
+    expect(shortString.decodeShortString(orderStatusBetween)).toBe("FULFILLED");
+  }, 60000);
 });
